Fist working version
This version contains:
  Buttons:
    * Start
    * Stop
    * Reset altitude
  Text field:
     * Start presure    [hPa]
     * Current presure  [hPa]
     * Current altitude [m]
     * Hello text

Start presure value is an average from 15 consecutive measurements.
Current presure value is an average from 5 last measurements.
This gives more stable and accurate results. A disadvantage is that
the reset start presure takes about 2 seconds.

The sensor is turned on only if the application is on foreground.
Everythime the application goes background the sensor is turned off.

Signed-off-by: Janusz Kozerski <janusz.kozerski@gmail.com>